update eslint and switch to flat config
