--- dotcom-rendering/.eslintignore
@@ -1,11 +1,0 @@
-node_modules
-.eslintrc.js
-storybook-static/
-
-# build output
-dist
-
-# playwright
-test-results/
-playwright-report/
-playwright/.cache/
