+++ dotcom-rendering/package.json
@@ -35,7 +35,7 @@
 		"@guardian/cdk": "62.6.1",
 		"@guardian/commercial-core": "32.0.0",
 		"@guardian/core-web-vitals": "7.0.0",
-		"@guardian/eslint-config-typescript": "12.0.0",
+		"@guardian/eslint-config": "12.0.1",
 		"@guardian/identity-auth": "6.0.1",
 		"@guardian/identity-auth-frontend": "8.1.0",
 		"@guardian/libs": "31.0.1",
@@ -89,6 +89,7 @@
 		"@types/webpack-env": "1.18.8",
 		"@types/webpack-node-externals": "3.0.4",
 		"@types/youtube": "0.0.50",
+		"@typescript-eslint/eslint-plugin": "8.22.0",
 		"ajv": "8.18.0",
 		"ajv-formats": "2.1.1",
 		"aws-cdk": "2.1115.1",
@@ -104,16 +105,15 @@
 		"css-loader": "7.1.2",
 		"dompurify": "3.2.4",
 		"dynamic-import-polyfill": "0.1.1",
-		"eslint": "8.57.1",
+		"eslint": "9.39.1",
 		"eslint-plugin-custom-elements": "0.0.8",
-		"eslint-plugin-jsx-a11y": "6.7.1",
+		"eslint-plugin-jsx-a11y": "6.10.2",
 		"eslint-plugin-jsx-expressions": "1.3.1",
 		"eslint-plugin-mocha": "10.1.0",
-		"eslint-plugin-react": "7.33.2",
-		"eslint-plugin-react-hooks": "4.6.0",
+		"eslint-plugin-react": "7.37.2",
+		"eslint-plugin-react-hooks": "5.1.0",
 		"eslint-plugin-ssr-friendly": "1.3.0",
-		"eslint-plugin-storybook": "10.2.13",
-		"eslint-plugin-unicorn": "48.0.1",
+		"eslint-plugin-unicorn": "64.0.0",
 		"eslint-stats": "1.0.1",
 		"express": "5.2.1",
 		"find": "0.3.0",
